+++ example/main.dart
@@ -9,7 +9,6 @@ void main() {
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -24,59 +23,12 @@ class MyApp extends StatelessWidget {
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
                 CustomImage(
-                  /// Set the path or URL for the image, accepting [String] types with extensions such as
-                  /// [jpg], [jpeg], [png], [gif], [webp], [svg]
-                  'https://pbs.twimg.com/media/FnM8-jFWYAIN2X7.jpg',
-
-                  /// set type image
-                  /// [ImageType.file],
-                  /// [ImageType.network],
-                  /// [ImageType.asset]
-                  type: ImageType.network,
-
-                  /// if you want to use onTap, you can use it like this
-                  onTap: () {
-                    /// implement your code here
-                  },
-
-                  /// set height and width for image
+                  'https://img.freepik.com/foto-gratis/pintura-lago-montana-montana-al-fondo_188544-9126.jpg',
                   height: 250,
                   width: 250,
-
-                  /// set borderRadius
-                  /// [BorderRadius.zero],
-                  /// [BorderRadius.circular],
-                  /// [BorderRadius.only],
-                  /// [BorderRadius.vertical],
-                  /// [BorderRadius.horizontal]
                   borderRadius: BorderRadius.circular(50),
-
-                  /// set color for image if you want
-                  color: Colors.transparent,
-
-                  /// set fit for image
-                  /// [FilterQuality.low],
-                  /// [FilterQuality.medium],
-                  /// [FilterQuality.high]
                   filterQuality: FilterQuality.high,
-
-                  /// set fit for image
-                  /// [BoxFit.contain],
-                  /// [BoxFit.cover],
-                  /// [BoxFit.fill],
-                  /// [BoxFit.fitHeight],
-                  /// [BoxFit.fitWidth],
-                  /// [BoxFit.none],
-                  /// [BoxFit.scaleDown],
-                  /// [BoxFit.fitHeight],
-                  /// [BoxFit.fitWidth],
                   fit: BoxFit.cover,
-
-                  /// set pathNoImage for a local image
-                  pathNoImage: 'assets/no_image.png',
-
-                  /// set pathLoading for a local image
-                  pathLoading: 'assets/loading.gif',
                 ),
               ],
             ),